TMDM-3357: Lock decorator does not display correctly on items in MDM repository view
Upon consultation with Haibo about the code, I have deleted the part that attempts to lock the item, as the item is locked elsewhere, and locking it again may cause error. However I have left in the part that causes a refresh when the editor is opened, thus making it more likely that the lock decorator appears.

Consulting with Haibo, the reason the lock decorator sometimes does not appear is due to the GUI not refreshing. To truly figure out everywhere a refresh might be missing might take a long time, and may not be warranted by the severity of the problem.

git-svn-id: http://talendforge.org/svn/tom/trunk@77956 f6f1c999-d317-4740-80b0-e6d1abc6f99e