merge r79216 from trunk to branch 5.0 TMDM-3264: Multiple language messages are not handled correctly in various parts of Studio and WebUI
This is in response to the report from QA that Validation Rule message parsing failed (#1, from bowang).

This occurs because the new code is no longer calling the code that I previously checked in to escape special characters like ] in studio. I have changed it back so that the method to escape characters is called again. Specifically, I have changed DescAnnotationListener's widgetSelected() method to call escapeMultiLanguageString() again, which was a method that was added in a previous commit for this issue.


git-svn-id: http://talendforge.org/svn/tom/branches/branch-5_0@79226 f6f1c999-d317-4740-80b0-e6d1abc6f99e